movemapp_client/tests/requirements.tests.txt
carto==1.11.3
certifi==2022.12.7
charset-normalizer==3.1.0
future==0.18.3
idna==3.4
pyrestcli==0.6.11
python-dateutil==2.8.2
python-dotenv==1.0.0
requests==2.28.2
six==1.16.0
urllib3==1.26.15
